Move type and parser exports out of route.ts
